refactor: construction crud
