+++ src/constructions/constructions.service.ts
@@ -15,13 +15,23 @@ export class ConstructionsService {
         description: createConstructionDto.description,
         date: new Date(createConstructionDto.date),
         hour: createConstructionDto.hour,
-        job_type_id: createConstructionDto.type,
+        job_type_id: createConstructionDto.job_type_id,
         categories: {
           connect: createConstructionDto.categories,
         },
+        items: {
+          create: createConstructionDto.items,
+        },
       },
       include: {
+        client: true,
         categories: true,
+        job_type: true,
+        items: {
+          include: {
+            item: true,
+          },
+        },
       },
     });
   }
@@ -44,6 +54,13 @@ export class ConstructionsService {
       include: {
         client: true,
         categories: true,
+        job_type: true,
+        items: {
+          include: {
+            item: true,
+          },
+        },
+        offers: true,
       },
     });
   }
@@ -54,6 +71,13 @@ export class ConstructionsService {
       include: {
         client: true,
         categories: true,
+        job_type: true,
+        items: {
+          include: {
+            item: true,
+          },
+        },
+        offers: true,
       },
     });
   }
@@ -76,6 +100,17 @@ export class ConstructionsService {
         },
         HttpStatus.BAD_REQUEST,
       );
+    if (updateConstructionDto.items) {
+      await this.prisma.construction.update({
+        where: { id },
+        data: {
+          items: {
+            deleteMany: {},
+            create: updateConstructionDto.items,
+          },
+        },
+      });
+    }
     return await this.prisma.construction.update({
       where: { id },
       data: {
@@ -84,7 +119,7 @@ export class ConstructionsService {
           ? new Date(updateConstructionDto.date)
           : updateConstructionDto.date,
         hour: updateConstructionDto.hour,
-        job_type_id: updateConstructionDto.type,
+        job_type_id: updateConstructionDto.job_type_id,
         status: updateConstructionDto.status,
         accepted_at: updateConstructionDto.accepted_at
           ? new Date(updateConstructionDto.accepted_at)
@@ -94,7 +129,14 @@ export class ConstructionsService {
         },
       },
       include: {
+        client: true,
         categories: true,
+        job_type: true,
+        items: {
+          include: {
+            item: true,
+          },
+        },
       },
     });
   }
