+++ src/constructions/decorators/construction-exists.decorator.ts
@@ -1,5 +1,5 @@
 import { registerDecorator, ValidationOptions } from 'class-validator';
-import { ConstructionExistsConstraint } from '../dto/validators/construction-exists.validator';
+import { ConstructionExistsConstraint } from '../validators/construction-exists.validator';
 
 export function ConstructionExists(validationOptions?: ValidationOptions) {
   return function (object: object, propertyName: string) {
